refactor perft fix cc1: warning: unrecognized command line option ‘-Wno-unused-command-line-argument’ and a few changes
