+++ src/makefile
@@ -2,7 +2,7 @@
 _THIS       := $(realpath $(dir $(abspath $(lastword $(MAKEFILE_LIST)))))
 _ROOT       := $(_THIS)
 TARGET      := Potential
-WARNINGS    = -Wall -Wextra -Wno-unused-command-line-argument -Wno-unused-variable
+WARNINGS    = -Wall -Wextra -Wno-unused-variable
 CFLAGS      := -funroll-loops -fomit-frame-pointer -Ofast -flto -fno-exceptions -DIS_64BIT -DNDEBUG $(WARNINGS)
 NATIVE      = -march=native
 AVX2FLAGS   = -DUSE_AVX2 -DUSE_SIMD -mavx2 -mbmi
